pyproject.toml
[project]
name = "adk-sample"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "google-adk>=1.18.0",
    "litellm>=1.80.0",
]
